pyproject.toml
[project]
name = "desktop-mcp"
version = "0.1.0"
description = "A vision-powered desktop automation library and MCP server. Control any GUI app using natural language via local or remote multimodal LLMs."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
keywords = ["mcp", "automation", "desktop", "gui", "vision", "llm", "pyautogui"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Topic :: Desktop Environment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pillow>=12.1.1",
    "pyautogui>=0.9.54",
    "pyperclip>=1.9.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
gui = ["flet[all]>=0.81.0"]

[project.scripts]
mcp_server = "mcp_server:main_sync"

[project.urls]
Repository = "https://github.com/snowmerak/desktop-mcp"
